edtech_platform/requirements.txt
aniso8601==7.0.0
asgiref==3.3.4
click==8.0.1
croniter==1.0.13
Django==3.2.4
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-rq==2.4.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
importlib-metadata==4.5.0
Markdown==3.3.4
promise==2.3
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
rq==1.8.1
rq-scheduler==0.11.0
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.10.0.0
zipp==3.4.1
